refactor: organize shared algorithm components
