+++ anomavision/algorithm/common/feature_extraction.py
