+++ anomavision/algorithm/patchcore/patchcore.py
@@ -12,7 +12,7 @@ from typing import Dict, List, Optional, Tuple
 import torch
 import torch.nn.functional as F
 
-from ...feature_extraction import ResnetEmbeddingsExtractor
+from ..common.feature_extraction import ResnetEmbeddingsExtractor
 
 
 class PatchCore(torch.nn.Module):
