+++ anomavision/mahalanobis/__init__.py
@@ -1,0 +1,5 @@
+"""Backward-compatible imports for the shared Mahalanobis distance module."""
+
+from ..algorithm.common.mahalanobis import MahalanobisDistance
+
+__all__ = ["MahalanobisDistance"]
